requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
boto3==1.20.0
botocore==1.23.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
docopt==0.6.2
idna==3.3
iniconfig==1.1.1
jmespath==0.10.0
packaging==21.2
pipreqs==0.4.11
pluggy==1.0.0
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
yaml-1.3==0.1.0
yarg==0.1.9
